+++ cli.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 import * as program from 'commander'
-import { error } from './utils'
+import { error, LngLat } from './utils'
 
 const geocoder = require('.')
 delete geocoder.default
@@ -10,6 +10,9 @@ interface CLI extends program.ICommand {
   limit?: number
   location?: string
   provider?: string
+  nearest?: LngLat
+  in?: Array<string>
+  distance?: number
 }
 
 function customHelp() {
@@ -30,9 +33,10 @@ program
   .usage('[options] <location>')
   .description('Geocoding results according to the GeoJSON specification')
   .option('-p, --provider [string="bing"]', 'Geocoding Provider')
-  .option('--limit [number]', 'Limit the results')
+  .option('--limit [number]', 'Limit the results', value => Number(value))
   .option('--nearest [LngLat]', 'Nearest result from a given LngLat', value => JSON.parse(value))
-  .option('--distance [number]', 'Maximum distance of nearest results')
+  .option('--in [Array<string>]', 'Filter by place=*', value => JSON.parse(value))
+  .option('--distance [number]', 'Maximum distance of nearest results', value => Number(value))
   .on('--help', customHelp)
   .parse(process.argv)